refactor query parameter and uri parsing code out of rest_json and rest_xml into rest_request_generator
